requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.12
starlette==0.41.3
browser-use==0.1.40
playwright==1.49.0
langchain-google-genai
python-dotenv==1.0.1
